// src/interfaces/player.tsx
export interface PlayerProps {
  id: string
  firstName: string
  lastName: string
  image: string
  attack: string
  defense: string
  skills: string
  idAuthor: string
  idPosition: string
}

export interface CardPlayerEditProps {
  idPlayer: string
  firstName: string
  lastName: string
  image: string
  attack: string
  defense: string
  skills: string
  idPosition: string
}